requirements.txt
Kivy==2.0.0rc3
Kivy-Garden==0.1.4
Pillow==7.2.0
Pygments==2.7.1
certifi==2020.6.20
chardet==3.0.4
docutils==0.16
idna==2.10
kivy-deps.angle==0.2.0
kivy-deps.glew==0.2.0
kivy-deps.sdl2==0.2.0
pip==20.2.3
requests==2.24.0
sdl2-lib==2.0.3
setuptools==47.1.0
urllib3==1.25.10
plyer==1.4.3
pytesseract==0.3.6
opencv-python==4.4.0.44
Cython==0.29.21
Jinja2==2.11.2
Kivy-examples==2.0.0rc3
MarkupSafe==1.1.1
appdirs==1.4.4
bs4==0.0.1
beautifulsoup4==4.9.3
autoconf==0.6.2
colorama==0.4.3
importlib-metadata==2.0.0
pep517==0.6.0
pip==20.2.3
pypiwin32==223
python-for-android==2020.6.2
pytoml==0.1.21
pywin32==228
sh==1.14.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.10
wheel==0.35.1
zipp==3.3.0

